requirements.txt
# FastAPI and Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0

# Scheduling
apscheduler==3.10.4

# Environment
python-dotenv==1.0.0

# Vector embeddings and Google Cloud AI
pgvector==0.2.5
google-cloud-aiplatform==1.40.0
google-cloud-storage==2.14.0
